finished kernel refactor
